+++ src-tauri/src/git/git_project_state.rs
@@ -1,0 +1,8 @@
+use serde::{Serialize, Deserialize};
+
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum GitProjectState {
+    Valid,
+    Invalid,
+}
