+++ src-tauri/src/git/mod.rs
@@ -2,3 +2,6 @@ pub mod git_files;
 pub mod git_folders;
 pub mod git_project;
 pub mod project_folder;
+pub mod git_commit;
+pub mod git_branch;
+pub mod git_project_state;
